Rename ServerAdaptor to Adaptor
Rename IOServerAdaptor to IOAdaptor
